Add some more scripts

+++ anml_from_snort.sh
@@ -1,9 +1,15 @@
-#!/bin/bash
+#!/bin/zsh
 set -eu
 
-if [[ $# -ne 2 ]] || $1 == "-h" ]]; then
-	echo "Usage: $0 <snort rules file> <output folder>"
+typeset -a leave_regex
+zparseopts -D -E -leave-regex=leave_regex
+
+if [[ $# -lt 2 ]] || [[ $1 == "-h" ]]; then
+	echo "Usage: $0 <snort rules file> <output folder> [flags for python]"
 	echo "Produces a number of ANML files that are the regular expressions for each distinct group"
+	echo "--leave-regex: leave the regex files"
+	echo "Python flags are:"
+	python regex_group_extractor.py --help
 	exit 1
 fi
 
@@ -18,5 +24,7 @@ for f in $output/*; do
 		mv automata_0.anml $f.anml
 		rm $f.mnrl
 	fi
-	rm $f
+	if [[ ${#leave_regex} -eq 0 ]]; then
+		rm $f
+	fi
 done
